+++ packages/kit/src/core/http/index.js
@@ -1,1 +1,2 @@
-export { get_body } from './get_body/index.js';
+export { get_raw_body } from './get_raw_body.js';
+export { parse_body } from './parse_body/index.js';
