+++ packages/kit/src/runtime/server/page/load_node.js
@@ -147,7 +147,9 @@ export async function load_node({
 								method: opts.method || 'GET',
 								headers,
 								path: resolved,
-								body: /** @type {any} */ (opts.body),
+								// TODO per https://developer.mozilla.org/en-US/docs/Web/API/Request/Request, this can be a
+								// Blob, BufferSource, FormData, URLSearchParams, USVString, or ReadableStream object
+								rawBody: opts.body,
 								query: new URLSearchParams(parsed.query || '')
 							},
 							options,
